Change the project name to the app name
