+++ app/src/androidTest/java/com/example/vibe/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import android.content.Context;
