+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.loginpage">
+    package="com.example.vibe">
 
     <application
         android:allowBackup="true"
@@ -14,19 +14,19 @@
         android:theme="@style/Theme.LoginPage"
         tools:targetApi="31">
         <activity
-            android:name=".Settings_View"
+            android:name="com.example.vibe.Settings_View"
             android:exported="false" />
         <activity
-            android:name=".conversation_view"
+            android:name="com.example.vibe.conversation_view"
             android:exported="false" />
         <activity
-            android:name=".login"
+            android:name="com.example.vibe.login"
             android:exported="false" />
         <activity
-            android:name=".register"
+            android:name="com.example.vibe.register"
             android:exported="false" />
         <activity
-            android:name=".SplashScrren"
+            android:name="com.example.vibe.SplashScrren"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -35,7 +35,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.vibe.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
