+++ app/src/main/java/com/example/vibe/FriendSearch.java
@@ -1,4 +1,4 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import android.os.Bundle;
