+++ app/src/main/java/com/example/vibe/SplashScrren.java
@@ -1,14 +1,18 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class conversation_view extends AppCompatActivity {
+public class SplashScrren extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_conversation_view);
+        setContentView(R.layout.activity_splash_scrren);
+
+
+
+
     }
 }
