+++ app/src/main/java/com/example/vibe/conversation_view.java
@@ -1,14 +1,14 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class MainActivity extends AppCompatActivity {
+public class conversation_view extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_conversation_view);
     }
 }
