+++ app/src/main/java/com/example/vibe/register.java
@@ -1,17 +1,18 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
 
-public class SplashScrren extends AppCompatActivity {
+public class register extends AppCompatActivity {
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash_scrren);
+        setContentView(R.layout.activity_register);
+
+
