+++ app/src/test/java/com/example/vibe/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.loginpage;
+package com.example.vibe;
 
 import org.junit.Test;
